extract the common training funcs
